requirements.txt
torch
torchvision
opencv-python
pandas
scikit-learn
tqdm
matplotlib
seaborn
pillow
# mediapipe
fastapi
uvicorn
numpy